Now use uint8_t to store qualities -> massive reduction of memory footprint.

+++ TDistanceEstimator.cpp
@@ -213,7 +213,7 @@ void TEMforDistanceEstimation::calculateDistance(){
 		distance += phi[i] * distanceWeight[i];
 }
 
-void TEMforDistanceEstimation::guessPi(std::vector<int*> & genoQual1, std::vector<int*> & genoQual2){
+void TEMforDistanceEstimation::guessPi(std::vector<uint8_t*> & genoQual1, std::vector<uint8_t*> & genoQual2){
 	//check sizes are equal
 	if(genoQual1.size() != genoQual2.size())
 		throw "Provided genotype quality vectors are of different size in TEMforDistanceEstimation::guessPi!";
@@ -223,8 +223,8 @@ void TEMforDistanceEstimation::guessPi(std::vector<int*> & genoQual1, std::vecto
 	double sum1, sum2;
 
 	//now loop over sites
-	std::vector<int*>::iterator it1 = genoQual1.begin();
-	std::vector<int*>::iterator it2 = genoQual2.begin();
+	std::vector<uint8_t*>::iterator it1 = genoQual1.begin();
+	std::vector<uint8_t*>::iterator it2 = genoQual2.begin();
 	for(; it1 != genoQual1.end(); ++it1, ++it2){
 		sum1 = 0.0; sum2 = 0.0;
 		for(int i=0; i<10; ++i){
@@ -245,7 +245,7 @@ void TEMforDistanceEstimation::guessPi(std::vector<int*> & genoQual1, std::vecto
 	pi.normalize();
 }
 
-void TEMforDistanceEstimation::guessPhi(std::vector<int*> & genoQual1, std::vector<int*> & genoQual2){
+void TEMforDistanceEstimation::guessPhi(std::vector<uint8_t*> & genoQual1, std::vector<uint8_t*> & genoQual2){
 	//check sizes are equal
 	if(genoQual1.size() != genoQual2.size())
 		throw "Provided genotype quality vectors are of different size in TEMforDistanceEstimation::guessPhi!";
@@ -256,8 +256,8 @@ void TEMforDistanceEstimation::guessPhi(std::vector<int*> & genoQual1, std::vect
 
 	//now loop over sites and add posterior probs
 	double sum1, sum2;
-	std::vector<int*>::iterator it1 = genoQual1.begin();
-	std::vector<int*>::iterator it2 = genoQual2.begin();
+	std::vector<uint8_t*>::iterator it1 = genoQual1.begin();
+	std::vector<uint8_t*>::iterator it2 = genoQual2.begin();
 	for(; it1 != genoQual1.end(); ++it1, ++it2){
 		sum1 = 0.0; sum2 = 0.0;
 		for(int i=0; i<10; ++i){
@@ -463,7 +463,7 @@ void TEMforDistanceEstimation::fill_P_g_given_phi_pi(double* thesePhi, TBaseFreq
 	probGeno[GT][AC] = tmp;
 }
 
-bool TEMforDistanceEstimation::estimatePhiWithEM(std::vector<int*> & genoQual1, std::vector<int*> & genoQual2){
+bool TEMforDistanceEstimation::estimatePhiWithEM(std::vector<uint8_t*> & genoQual1, std::vector<uint8_t*> & genoQual2){
 	//prepare estimates
 	logfile->listFlush("Estimating initial base frequencies pi ...");
 	guessPi(genoQual1, genoQual2);
@@ -498,8 +498,8 @@ bool TEMforDistanceEstimation::estimatePhiWithEM(std::vector<int*> & genoQual1, 
 		}
 
 		//loop across loci to calculate P_G
-		std::vector<int*>::iterator it1 = genoQual1.begin();
-		std::vector<int*>::iterator it2 = genoQual2.begin();
+		std::vector<uint8_t*>::iterator it1 = genoQual1.begin();
+		std::vector<uint8_t*>::iterator it2 = genoQual2.begin();
 		for(; it1 != genoQual1.end(); ++it1, ++it2){
 			//calculate P_G per site
 			double sum = 0.0;
@@ -748,7 +748,7 @@ bool TDistanceEstimator::advance(TGlfReader & g1, TGlfReader & g2){
 	return(moveToNextCommonChr(g1, g2));
 };
 
-void TDistanceEstimator::readCommonSites(std::vector<int*> & genoQual1, std::vector<int*> & genoQual2, TGlfReader & g1, TGlfReader & g2){
+void TDistanceEstimator::readCommonSites(std::vector<uint8_t*> & genoQual1, std::vector<uint8_t*> & genoQual2, TGlfReader & g1, TGlfReader & g2){
 	//parse GLFs. Only keep sites where both individuals have data!
 
 	//rewind GLFs
@@ -759,8 +759,8 @@ void TDistanceEstimator::readCommonSites(std::vector<int*> & genoQual1, std::vec
 	while(advance(g1, g2)){
 		if(g2.position == g1.position){
 			//add data
-			genoQual1.push_back(new int[10]);
-			genoQual2.push_back(new int[10]);
+			genoQual1.push_back(new uint8_t[10]);
+			genoQual2.push_back(new uint8_t[10]);
 			g1.fillGenotypeQualities(*genoQual1.rbegin());
 			g2.fillGenotypeQualities(*genoQual2.rbegin());
 		}
@@ -770,7 +770,7 @@ void TDistanceEstimator::readCommonSites(std::vector<int*> & genoQual1, std::vec
 void TDistanceEstimator::estimateDistanceGenomeWide(TEMforDistanceEstimation & EM_object, TGlfReader & g1, TGlfReader & g2, gz::ogzstream & out){
 	//initialize storage for two windows
 	logfile->listFlush("Reading common sites ...");
-	std::vector<int*> genoQual1, genoQual2;
+	std::vector<uint8_t*> genoQual1, genoQual2;
 	readCommonSites(genoQual1, genoQual2, g1, g2);
 	logfile->done();
 	logfile->conclude("Read data for " + toString(genoQual1.size()) + " sites.");
@@ -788,9 +788,9 @@ void TDistanceEstimator::estimateDistanceGenomeWide(TEMforDistanceEstimation & E
 
 	//clean up memory
 	logfile->listFlush("Cleaning up memory ...");
-	for(std::vector<int*>::iterator it=genoQual1.begin(); it!=genoQual1.end(); ++it)
+	for(std::vector<uint8_t*>::iterator it=genoQual1.begin(); it!=genoQual1.end(); ++it)
 		delete[] *it;
-	for(std::vector<int*>::iterator it=genoQual2.begin(); it!=genoQual2.end(); ++it)
+	for(std::vector<uint8_t*>::iterator it=genoQual2.begin(); it!=genoQual2.end(); ++it)
 		delete[] *it;
 	genoQual1.clear();
 	genoQual2.clear();
@@ -835,12 +835,12 @@ void TDistanceEstimator::estimateDistanceInWindows(TEMforDistanceEstimation & EM
 
 	//initialize storage for two windows
 	//TODO: share across pairs? Do all pairs at once?
-	std::vector<int*> genoQual1, genoQual2;
+	std::vector<uint8_t*> genoQual1, genoQual2;
 	genoQual1.resize(windowLen);
 	genoQual2.resize(windowLen);
 	for(int i=0; i<windowLen; ++i){
-		genoQual1[i] = new int[10];
-		genoQual2[i] = new int[10];
+		genoQual1[i] = new uint8_t[10];
+		genoQual2[i] = new uint8_t[10];
 	}
 
 	//open output file
